Parse any token
The alt! macro use on line 47 highlights the real power of nom.
In previous steps, we made a bunch of little, easy to read parser
that have one job. Now, with this macro invocation, we have a
declarative syntax that is basically just a list of valid tokens.

I added a few tests for this, just to prove it runs without error,
but most of its behavior is covered by the lower-level tests on
the individual parsers.