Parse vector of tokens
The goal of this step is to parse strings into vectors of
tokens. Again, here nom really shines, because it can just say,
"hey, this parser will expect a byte array to be many sequences
that match the underlying parsers." I thought this step would be harder.

I added tests that check the parsing of simple strings. I wanted to
have a test for multiple digit numeric tokens, because all the other
tokens will be single bytes in a normal string.