Implement shift
We're implementing a shift-reduce parser. We have two halves: a stack
of tokens we've already seen, and a slice of tokens we haven't seen
yet.

There are two basic operations: Shift, which takes the first item in
the unseen tokens and pushes it onto the stack, and reduce, which
simplifies the stack by applying operators to operands that are
already on the stack.

This commit is my first pass at implementing shift.